2nd Rendition of puzzle prototype
solution to validating connections by adding a reference to the previous tile, like a linked list and iterating through the list checking if its not null. Also I check for neighboring tiles by giving the tiles a X and Y Id and checking if its within range by 1.  The problem with this is that diagonals will still be considered a connection as well.